cxx-qt: rename cxxqtthread module to threading
